\name{equalizeSe}
\alias{equalizeSe}
\title{Forces standard errors (se) elements in a list of tcl variables to be equal}
\usage{
  equalizeSe(...)
}
\description{
  This function is called by pressing the "Equalize"
  sliders and overwrites individual standard error values
  in a list of tcl variables.
}
\keyword{internal}

