\name{equalizeT}
\alias{equalizeT}
\title{Forces effect size (T) elements in a list of tcl variables to be equal}
\usage{
  equalizeT(...)
}
\description{
  This function is called by pressing the "Equalize"
  sliders and overwrites individual effect size values in a
  list of tcl variables.
}
\keyword{internal}

