\name{OnOK}
\alias{OnOK}
\title{Main function to be called by most Tk buttons and sliders}
\usage{
  OnOK(...)
}
\description{
  Each time a button is pressed, a slider is activated or a
  new data set is loaded, OnOK() is called and causes the
  graphics window as well as the summary statistics to be
  updated.
}
\keyword{internal}

