\name{plotForest}
\alias{plotForest}
\title{Creates a simple forest plot without}
\usage{
  plotForest(T, se, index, ll, ul, mylinewidth)
}
\arguments{
  \item{T}{}

  \item{se}{}

  \item{index}{}

  \item{ll}{}

  \item{ul}{}

  \item{mylinewidth}{}
}
\description{
  Creates a simple forest plot without
}
\keyword{internal}

