\name{plotMeta}
\alias{plotMeta}
\title{Open new plot window and creates forest and density plot}
\usage{
  plotMeta(T, se, k, studlab, myfontsize, mylinewidth)
}
\arguments{
  \item{T}{numeric vector of effect sizes}

  \item{se}{numeric vector of standard errors}

  \item{k}{length of T}

  \item{studlab}{character vector of study labels}

  \item{myfontsize}{font size (cex)}

  \item{mylinewidth}{line width (lwd)}
}
\description{
  Open new plot window and creates forest and density plot
}
\keyword{internal}

