\name{tclVarList2vector}
\alias{tclVarList2vector}
\title{Convert list of tclVar variables to vector}
\usage{
  tclVarList2vector(x)
}
\arguments{
  \item{x}{a list of tclVar() objects}
}
\description{
  Takes a list of tclVar objects and returns a numeric
  vector or a character vector which then can be processed
  by other functions.
}
\keyword{internal}

