\docType{package}
\name{tuma}
\alias{package-tuma}
\alias{tuma}
\alias{tuma-package}
\title{tuma: A graphical user interface for teaching univariate meta-analysis}
\description{
  This package provides a graphical user interface (GUI) to
  visualize univariate fixed effect (FEM) and random
  effects model (REM) meta-analysis. By interactively
  manipulating effect sizes and standard errors, it is
  possible to better understand their relationship with
  summary (FEM and REM) and heterogeneity related
  statistics (Q, I^2 and tau^2). Additionally, two plots
  are created which show a typical forest plot as well as a
  density plot with N(hat{theta}, tau^2). These plots also
  change if the individual effects sizes or standard errors
  are manipulated. This work was supported by a fellowship
  within the "Postdoc-Programme" of the German Academic
  Exchange Service (DAAD).
}
\details{
  The tuma GUI can be started by typing \code{tumaGUI()}.
}

